Improve conditioning of the Kalman filter design
Add a little process noise to all states in during the Kalman filter
design.  This ensures that the process noise covariance matrix is
strictly positive definite, which together with the measurement noise
covariance being strictly positive definite makes a set of sufficient
(but not necessary) conditions for the solution of the Ricatti equation
to exist.

Matlab's `lqr` function is used because it allows noise to be added to
the states more easily than the `kalman` function. This is allowed
because of the duality of LQR and LQE problems.

The numerical values for the covariance are largely unchanged: the
measurement noise covariance is kept to 1e-6, the process noise entering
the turbulence channel has 1e-2 covariance and now process noise with
1e-8*eye(n) covariance is added to the states.